Fixing the login button crash and refactored and cleaned up code
