+++ Samples/BasicApp-Win81-Universal/FacebookConfig.xml
@@ -1,0 +1,5 @@
+﻿<?xml version="1.0" encoding="utf-8" ?>
+<Extensions>
+    <Protocol Name="msft-8ea05b63db9e44f886940975945fca72" NavUriFragment="encodedLaunchUri=%s" TaskID="_default" />
+    <Facebook AppId="540541885996234" />
+</Extensions>
