+++ Samples/BasicApp-Win81-Universal/MainPage.xaml
@@ -9,9 +9,10 @@
     mc:Ignorable="d">
 
     <Grid Background="{ThemeResource ApplicationPageBackgroundThemeBrush}">
+        <Controls:LoginButton HorizontalAlignment="Left" Margin="1092,668,0,0" VerticalAlignment="Top"/>
+        
+        <Controls:FriendPicker HorizontalAlignment="Left" Margin="86,54,0,0" VerticalAlignment="Top" Height="598" Width="935"/>
 
-        <!--<Controls:LoginButton HorizontalAlignment="Left" Margin="1186,688,0,0" VerticalAlignment="Top"/>-->
-        <Controls:FriendPicker/>
 
     </Grid>
 </Page>
