+++ Samples/BasicApp-Win81-Universal/packages.config
@@ -1,0 +1,4 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="Facebook" version="6.6.0" targetFramework="win81" />
+</packages>
