+++ Samples/BasicAppWP81/MainPage.xaml
@@ -416,7 +416,6 @@
             <!--Login Button-->
             <facebookControls:LoginButton x:Name="loginButton" 
                 Grid.Row="2"
-                ApplicationId="540541885996234"
                 HorizontalAlignment="Right"
                 VerticalAlignment="Bottom"
                 UserInfoChanged="OnUserInfoChanged"
