+++ Source/Facebook.Client-Universal/Facebook.Client-Universal.Windows/Controls/LoginButton/LoginButton.xaml
@@ -33,7 +33,6 @@
             <Setter.Value>
                 <ControlTemplate TargetType="local:LoginButton">
                     <Button x:Name="PART_LoginButton"
-                            IsEnabled="{Binding RelativeSource={RelativeSource TemplatedParent}, Path=ApplicationId, Converter={StaticResource IsNullOrEmptyConverter}, ConverterParameter=True}"
                             Background="{TemplateBinding Background}" 
                             Foreground="{TemplateBinding Foreground}" 
                             BorderBrush="{TemplateBinding BorderBrush}"
