+++ Source/Facebook.Client-Universal/Facebook.Client-Universal.Windows/Facebook.Client-Universal.Windows.csproj
@@ -216,6 +216,9 @@
     <Compile Include="..\..\Facebook.Client\FacebookSessionCacheProvider.cs">
       <Link>FacebookSessionCacheProvider.cs</Link>
     </Compile>
+    <Compile Include="..\..\Facebook.Client\FacebookSessionLocalSettingsCacheProvider.cs">
+      <Link>FacebookSessionLocalSettingsCacheProvider.cs</Link>
+    </Compile>
     <Compile Include="..\..\Facebook.Client\GraphLocation.cs">
       <Link>GraphLocation.cs</Link>
     </Compile>
@@ -244,7 +247,6 @@
     <Compile Include="FacebookDialogPage.xaml.cs">
       <DependentUpon>FacebookDialogPage.xaml</DependentUpon>
     </Compile>
-    <Compile Include="FacebookSessionRoamingSettingsCacheProvider.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
   <ItemGroup>
